Switch order of verification
This isn't the final solution, but I think it's a step in the right
direction - we're getting false-positives on external link checks that
cause data verification to never happen. This commit ensure that we
check things that I have confidence in first and then the thing that I
have the least confidence in last.